Support for one-way offers combinations
