+++ package.json
@@ -18,6 +18,7 @@
     "csv-parser": "^2.3.2",
     "date-fns": "^2.10.0",
     "jquery": "^3.4.1",
+    "lz-string": "^1.4.4",
     "mongodb": "^3.5.5",
     "node-sass": "^4.13.1",
     "query-string": "^6.12.1",
