+++ src/utils/offer-utils.js
@@ -161,71 +161,4 @@ export function iataToAirportName(iata){
   }
 }
 
-export const expandSearchResult = (criteria, searchResults) => {
-  // Check if we have a roundtrip flight Search
-  let offers = searchResults.offers;
-  //console.log('criteria >> ', criteria);
-  
-  if(criteria.returnDate !== undefined) {
-    // Prepare to split inbound and outbound offers
-    const orgIata = criteria.origin.code;
-    
-    let inboundOffers = [];
-    let outboundOffers = [];
 
-    // Go through each offer
-    Object.keys(offers).forEach(offerId => {
-      // Extract the price plans
-      const pricePlans = Object.keys(offers[offerId].pricePlansReferences);
-
-      // If there is only one price plan with one flight, it is a one-way offer
-      if(pricePlans.length === 1 && 
-        offers[offerId].pricePlansReferences[pricePlans[0]].flights.length === 1) {
-          const flightKey = offers[offerId].pricePlansReferences[pricePlans[0]].flights[0];
-          const segmentKeys = searchResults.itineraries.combinations[flightKey];
-
-          // Add offer to outbound or inbound arrays
-          if(searchResults.itineraries.segments[segmentKeys[0]].origin.iataCode === orgIata) {
-            outboundOffers.push(offerId);
-          } else {
-            inboundOffers.push(offerId);
-          }
-          
-      }
-    });
-
-    // Create new combinations using the outbound and inboud
-    inboundOffers.forEach(inboundOfferId => {
-      outboundOffers.forEach(outboundOfferId => {
-        const offerKey = `${outboundOfferId},${inboundOfferId}`;
-        const inboundOffer = offers[inboundOfferId];
-        const outboundOffer = offers[outboundOfferId];
-        //console.log('Merging offers: ', JSON.stringify(inboundOffer), JSON.stringify(outboundOffer)); 
-        offers[offerKey] = {
-          offerItems: {
-            ...inboundOffer.offerItems,
-            ...outboundOffer.offerItems,
-          },
-          expiration: Date(inboundOffer.expiration) < Date(outboundOffer.expiration) ? inboundOffer.expiration : outboundOffer.expiration,
-          pricePlansReferences: {
-            ...inboundOffer.pricePlansReferences,
-            ...outboundOffer.pricePlansReferences,
-          },
-          price: {
-            currency: inboundOffer.price.currency,
-            public: Number(Number(inboundOffer.price.public)+Number(outboundOffer.price.public)).toFixed(2),
-            taxes: Number(Number(inboundOffer.price.taxes)+Number(outboundOffer.price.taxes)).toFixed(2),
-          }
-        }
-      });
-    });
-
-    // Remove duplicated keys
-    [...inboundOffers,  ...outboundOffers].forEach(offerId => {
-      delete(offers[offerId]);
-    });
-
-    searchResults.offers = offers;
-  }
-  return searchResults;
-};
