+++ src/utils/search.js
@@ -3,6 +3,7 @@ import {extendResponse} from "./flight-search-results-transformer";
 import offline_flight_results from "../data/sample_response_flights";
 import offline_hotels_results from "../data/sample_response_hotels";
 import {uiEvent} from "./events";
+import LZString from "lz-string"
 
 /**
  * Search for flights with a criteria provided as parameters.
@@ -66,14 +67,14 @@ export async function findHotels(criteria) {
 }
 
 export function retrieveSearchResultsFromLocalStorage(){
-    let searchResults = JSON.parse(localStorage.getItem('searchResults'));
-    console.log("retrieveSearchResultsFromLocalStorage",searchResults)
+    let searchResults = JSON.parse(LZString.decompressFromUTF16(localStorage.getItem('searchResults')));
+    //console.log("retrieveSearchResultsFromLocalStorage",searchResults)
     return searchResults;
 }
 export function clearSearchResultsInLocalStorage(){
-    localStorage.setItem('searchResults',{})
+    localStorage.setItem('searchResults',LZString.compressToUTF16({}));
 }
 export function storeSearchResultsInLocalStorage(searchResults){
-    console.log("storeSearchResultsInLocalStorage",searchResults)
-    localStorage.setItem('searchResults',JSON.stringify(searchResults))
+    //console.log("storeSearchResultsInLocalStorage",searchResults)
+    localStorage.setItem('searchResults', LZString.compressToUTF16(JSON.stringify(searchResults)));
 }
